Refactors files' names
